ai-service/requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
httpx
psycopg2-binary
python-jose
pydantic
